Move files and add new functionality to services
